adds wip audit timeout test, moves SlowTransport to separate file
